Restructure files, yes, again
