+++ vs_meetup.php
@@ -17,14 +17,18 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 
 if ( ! defined( 'VSMEET_TEMPLATE_DIR' ) ) {
-	define( 'VSMEET_TEMPLATE_DIR', dirname( __FILE__ ) . '/templates/' );
+	define( 'VSMEET_TEMPLATE_DIR', dirname( __FILE__ ) . '/includes/templates/' );
 }
 
-require_once( 'class-meetup-widgets-admin.php' );
-require_once( 'class-meetup-widget.php' );
-require_once( 'class-meetup-api-v3.php' );
-require_once( 'api/class-meetup-rest-events-controller.php' );
-require_once( 'api/class-meetup-rest-groups-controller.php' );
+require_once( 'includes/class-meetup-widgets-admin.php' );
+require_once( 'includes/class-meetup-widget.php' );
+require_once( 'includes/class-meetup-api-v3.php' );
+require_once( 'includes/api/class-meetup-rest-events-controller.php' );
+require_once( 'includes/api/class-meetup-rest-groups-controller.php' );
+require_once( 'includes/blocks/blocks.php' );
+require_once( 'includes/widgets/single.php' );
+require_once( 'includes/widgets/group-list.php' );
+require_once( 'includes/widgets/user-list.php' );
 
 /**
  * Initialize Meetup Widgets
@@ -39,25 +43,20 @@ function meetup_widgets_start() {
 
 } add_action( 'init', 'meetup_widgets_start' );
 
-require_once( 'widgets/single.php' );
 add_action(
 	'widgets_init', function() {
 		return register_widget( 'VsMeetSingleWidget' );
 	}
 );
 
-require_once( 'widgets/group-list.php' );
 add_action(
 	'widgets_init', function() {
 		return register_widget( 'VsMeetListWidget' );
 	}
 );
 
-require_once( 'widgets/user-list.php' );
 add_action(
 	'widgets_init', function() {
 		return register_widget( 'VsMeetUserListWidget' );
 	}
 );
-
-require_once( 'blocks/blocks.php' );
